+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/Components.java
@@ -925,7 +925,7 @@ public final class Components {
 	 * @return an indeterminate JProgressBar builder
 	 */
 	public static ProgressBarBuilder progressBar() {
-		return ProgressBarBuilder.builder(null);
+		return ProgressBarBuilder.builder();
 	}
 
 	/**