+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/progressbar/DefaultProgressBarBuilder.java
@@ -23,6 +23,7 @@ import is.codion.swing.common.ui.component.value.AbstractComponentValue;
 import is.codion.swing.common.ui.component.value.ComponentValue;
 
 import javax.swing.BoundedRangeModel;
+import javax.swing.DefaultBoundedRangeModel;
 import javax.swing.JProgressBar;
 
 final class DefaultProgressBarBuilder extends AbstractComponentBuilder<Integer, JProgressBar, ProgressBarBuilder> implements ProgressBarBuilder {
@@ -36,8 +37,8 @@ final class DefaultProgressBarBuilder extends AbstractComponentBuilder<Integer, 
 	private String string;
 
 	DefaultProgressBarBuilder(BoundedRangeModel boundedRangeModel) {
-		this.boundedRangeModel = boundedRangeModel;
 		this.indeterminate = boundedRangeModel == null;
+		this.boundedRangeModel = boundedRangeModel == null ? new DefaultBoundedRangeModel() : boundedRangeModel;
 	}
 
 	@Override